Fixed compaction and removed memtable synchronization

+++ codec/codec-test/src/test/java/io/joshworks/fstore/codec/test/CodecTest.java
@@ -192,7 +192,7 @@ public abstract class CodecTest {
 
             compressed.flip();
 
-            int compressedSize = compressed.remaining();
+            double compressedSize = compressed.remaining();
             compression += (dataSize - compressedSize) / 100;
